refactored tests to be more resemblant to reality even though they will be scrapped
