+++ src/integration/java/com/askoura/travis/TravisCIVerifierIntegrationTest.java
@@ -7,33 +7,25 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import com.askoura.helper.DBHelper;
-import com.askoura.library.config.Persistence;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = {Persistence.class})
+@ContextConfiguration(classes = {TestConf.class})
 public class TravisCIVerifierIntegrationTest {
 
     @Autowired
-    private JdbcTemplate jdbc;
-
     private DBHelper dbHelper = null;
-
-    @Before
-    public void initDbHelper() {
-        dbHelper = new DBHelper(jdbc);
-    }
     
+    @Autowired
+    private TravisCIVerifier travisCIVerifier;
 
     @Test
     public void returnWhatYouGet_givingTrue_expectingTrue() throws Exception {
-        // Given
-        TravisCIVerifier travisCIVerifier = new TravisCIVerifier(jdbc);
-        
         // When
         boolean result = travisCIVerifier.returnWhatYouGet(true);
         
@@ -43,9 +35,6 @@ public class TravisCIVerifierIntegrationTest {
 
     @Test
     public void onlyInIntegrationSum_1Plus2_return3() throws Exception {
-        // Given
-        TravisCIVerifier travisCIVerifier = new TravisCIVerifier(jdbc);
-
         // When
         int sum = travisCIVerifier.onlyInIntegrationSum(1, 2);
 
@@ -55,9 +44,6 @@ public class TravisCIVerifierIntegrationTest {
 
     @Test
     public void putSomeStuffInDb_getSomeStuffFromDB_inMemDBWorks() throws Exception {
-        // Given
-        TravisCIVerifier travisCIVerifier = new TravisCIVerifier(jdbc);
-
         // When
         travisCIVerifier.insertIntoTableA(1);
         travisCIVerifier.insertIntoTableA(2);
